Break out remaining commands
This commit finishes the clean-up of `pkg/cmd/pulumi`, breaking out the
`gen-completion` and `gen-markdown` commands.
