+++ pkg/cmd/pulumi/completion/gen.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2018, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package completion
 
 import (
 	"bytes"
@@ -26,8 +26,8 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
 )
 
-// newGenCompletionCmd returns a new command that, when run, generates a bash or zsh completion script for the CLI.
-func newGenCompletionCmd(root *cobra.Command) *cobra.Command {
+// NewGenCompletionCmd returns a new command that, when run, generates a bash or zsh completion script for the CLI.
+func NewGenCompletionCmd(root *cobra.Command) *cobra.Command {
 	return &cobra.Command{
 		Use:     "gen-completion <SHELL>",
 		Aliases: []string{"completion"},
